updated subgraph to use templates
